requirements.txt
asgiref==3.7.2
astroid==3.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crossplane==0.5.8
Django==4.1.12
django-countries==7.5.1
django-filter==23.3
django-guardian==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
interrogate==1.5.0
isort==5.12.0
itypes==1.2.0
jellyfish==1.0.1
Jinja2==3.1.2
jproperties==2.1.1
luhn==0.2.0
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
py==1.11.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
ruff==0.1.2
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
whispers==2.1.5
